pyproject.toml
[project]
name = "extract-entities-whatsapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.19",
    "langchain-openai>=0.3.6",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.2",
]
